FESTOU-82 - Ajustando appbar e separando widgets em arquivos
